Evaluate the chi2 and chi2 gradient at the same time
This can speed up the chi2 computation for complex functions which are
expensive to compute. The function value itself is needed anyway for
the gradient, so when computing the gradient the chi2 value comes for free.
This reduces the number of function evaluations by one third when fitting
with a function with gradient.