Add an interface to evaluate both value and gradient to IParametricGradFunctionMultiDim
This is another optimization for chi2 fitting with a gradient function. Until
now, the chi2 function was computing separately the function value and its
gradient. Since often it can be implemented more efficienctly if the two
are computed at the same time a corresponding interface has been added to the
generic paramatric function interface, and used in the chi2 function. The
implementation of the combined computation is optional and, if not provided,
the two values are computed separately.